fix(migration): address Task 1 code-review P1/P2 findings
Task 1 code-quality review returned PASS_WITH_CONCERNS. Accepted fixes:

- P1: UserChatbotPermission.ID uint64 → uint (align with gorm.Model
  convention; GORM maps uint → BIGINT UNSIGNED internally)
- P1: expand model doc comment to explicitly document "no UpdatedAt /
  DeletedAt, append-only whitelist" design decision
- P2-A: backfill rollback .sql switched from :placeholder (Oracle/psql)
  to @session variable syntax (MySQL 8.x compatible). Added SET @... at
  top so runbook can just edit timestamps there.
- P2-B: create migration created_at default changed from NULL to
  NOT NULL DEFAULT CURRENT_TIMESTAMP(3), aligning with existing
  action_log.sql convention.

Declined: P1-A (COLLATE not explicit) — existing migrations follow the
same "CHARSET=utf8mb4 without explicit COLLATE" style; consistency with
project history outweighs forward-looking concern about future varchar
columns (current table has zero varchar).

Re-validated on dev DB: DROP + re-apply → SHOW CREATE TABLE confirms
new default timestamp. No prod data touched.

+++ internal/pkg/model/user_chatbot_permission.go
@@ -3,10 +3,14 @@ package model
 import "time"
 
 // UserChatbotPermission 子账号 chatbot 运行权限白名单（spec §3.1 / §3.3）。
-// 对称于 UserTemplatePermission（SOP 模板权限），但不使用软删除：
-// DELETE 直接真删行，语义上"0 记录 = deny-all"（default-deny）。
+//
+// 设计决策：
+//   - 对称于 UserTemplatePermission（SOP 模板权限），但**不嵌入 gorm.Model** —— 无 UpdatedAt / DeletedAt
+//   - 语义上是**append-only 白名单**：授权 = INSERT，撤权 = 物理 DELETE，永不 UPDATE
+//   - "0 记录 = deny-all"（default-deny），由 biz 层 HasChatbotPermission 判定
+//   - ID 类型用 uint（对齐 gorm.Model.ID 约定，GORM 内部映射到 SQL BIGINT UNSIGNED）
 type UserChatbotPermission struct {
-	ID        uint64    `gorm:"primaryKey;autoIncrement" json:"id"`
+	ID        uint      `gorm:"primaryKey;autoIncrement" json:"id"`
 	SubUserID uint      `gorm:"not null;uniqueIndex:uk_ucp_sub_chatbot" json:"sub_user_id"`
 	ChatbotID uint      `gorm:"not null;uniqueIndex:uk_ucp_sub_chatbot;index:idx_ucp_chatbot" json:"chatbot_id"`
 	CreatedAt time.Time `json:"created_at"`
